+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Enhancements
 
-* Add Mg/Ca seawater correction to `predict_mgca` with `sw_age` arg for Deep Time prediction (Issue #3)
+* Add Mg/Ca seawater correction to `predict_mgca` with `sw_age` arg for Deep Time prediction (Issue #3).
+* Vectorize `predict_mgca`.
 
 
 ## Bug fixes