+++ NEWS.md
@@ -1,3 +1,11 @@
+# v0.0.1a3
+
+## Enhancements
+
+
+## Bug fixes
+
+
 # v0.0.1a2
 
 ## Enhancements